requirements.txt
streamlit
tensorflow
pillow
numpy
matplotlib
streamlit-drawable-canvas
